Move part of debugutils.hpp to testutils.cpp
