+++ include/tlapack/plugins/debugutils.hpp
@@ -16,7 +16,6 @@
 #include <iostream>
 #include <sstream>
 
-#include "tlapack/base/legacyArray.hpp"
 #include "tlapack/base/utils.hpp"
 
 namespace tlapack {
@@ -38,49 +37,6 @@ void print_matrix(const matrix_t& A)
         for (idx_t j = 0; j < n; ++j)
             std::cout << std::setw(16) << A(i, j) << " ";
     }
-}
-
-//
-// GDB doesn't handle templates well, so we explicitly define some versions of
-// the functions for common template arguments
-//
-void print_matrix_r(const legacyMatrix<float, size_t, Layout::ColMajor>& A)
-{
-    print_matrix(A);
-}
-void print_matrix_d(const legacyMatrix<double, size_t, Layout::ColMajor>& A)
-{
-    print_matrix(A);
-}
-void print_matrix_c(
-    const legacyMatrix<std::complex<float>, size_t, Layout::ColMajor>& A)
-{
-    print_matrix(A);
-}
-void print_matrix_z(
-    const legacyMatrix<std::complex<double>, size_t, Layout::ColMajor>& A)
-{
-    print_matrix(A);
-}
-void print_rowmajormatrix_r(
-    const legacyMatrix<float, size_t, Layout::RowMajor>& A)
-{
-    print_matrix(A);
-}
-void print_rowmajormatrix_d(
-    const legacyMatrix<double, size_t, Layout::RowMajor>& A)
-{
-    print_matrix(A);
-}
-void print_rowmajormatrix_c(
-    const legacyMatrix<std::complex<float>, size_t, Layout::RowMajor>& A)
-{
-    print_matrix(A);
-}
-void print_rowmajormatrix_z(
-    const legacyMatrix<std::complex<double>, size_t, Layout::RowMajor>& A)
-{
-    print_matrix(A);
 }
 
 /**
@@ -183,51 +139,6 @@ std::string visualize_matrix(const matrix_t& A)
         return visualize_matrix_text(A);
     else
         return visualize_matrix_table(A);
-}
-
-//
-// GDB doesn't handle templates well, so we explicitly define some versions of
-// the functions for common template arguments
-//
-std::string visualize_matrix_r(
-    const legacyMatrix<float, size_t, Layout::ColMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_matrix_d(
-    const legacyMatrix<double, size_t, Layout::ColMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_matrix_c(
-    const legacyMatrix<std::complex<float>, size_t, Layout::ColMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_matrix_z(
-    const legacyMatrix<std::complex<double>, size_t, Layout::ColMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_rowmajormatrix_r(
-    const legacyMatrix<float, size_t, Layout::RowMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_rowmajormatrix_d(
-    const legacyMatrix<double, size_t, Layout::RowMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_rowmajormatrix_c(
-    const legacyMatrix<std::complex<float>, size_t, Layout::RowMajor>& A)
-{
-    return visualize_matrix(A);
-}
-std::string visualize_rowmajormatrix_z(
-    const legacyMatrix<std::complex<double>, size_t, Layout::RowMajor>& A)
-{
-    return visualize_matrix(A);
 }
 
 }  // namespace tlapack
