+++ test/include/testutils.hpp
@@ -16,6 +16,9 @@
 // Definitions
 #include "testdefinitions.hpp"
 
+// Plugin for debug
+#include <tlapack/plugins/debugutils.hpp>
+
 // <T>LAPACK
 #include <tlapack/base/utils.hpp>
 #include <tlapack/blas/gemm.hpp>
@@ -205,6 +208,46 @@ real_type<type_t<matrix_t>> check_similarity_transform(matrix_t& A,
 
     return check_similarity_transform(A, Q, B, res, work);
 }
+
+//
+// GDB doesn't handle templates well, so we explicitly define some versions of
+// the functions for common template arguments
+//
+void print_matrix_r(const legacyMatrix<float, size_t, Layout::ColMajor>& A);
+void print_matrix_d(const legacyMatrix<double, size_t, Layout::ColMajor>& A);
+void print_matrix_c(
+    const legacyMatrix<std::complex<float>, size_t, Layout::ColMajor>& A);
+void print_matrix_z(
+    const legacyMatrix<std::complex<double>, size_t, Layout::ColMajor>& A);
+void print_rowmajormatrix_r(
+    const legacyMatrix<float, size_t, Layout::RowMajor>& A);
+void print_rowmajormatrix_d(
+    const legacyMatrix<double, size_t, Layout::RowMajor>& A);
+void print_rowmajormatrix_c(
+    const legacyMatrix<std::complex<float>, size_t, Layout::RowMajor>& A);
+void print_rowmajormatrix_z(
+    const legacyMatrix<std::complex<double>, size_t, Layout::RowMajor>& A);
+
+//
+// GDB doesn't handle templates well, so we explicitly define some versions of
+// the functions for common template arguments
+//
+std::string visualize_matrix_r(
+    const legacyMatrix<float, size_t, Layout::ColMajor>& A);
+std::string visualize_matrix_d(
+    const legacyMatrix<double, size_t, Layout::ColMajor>& A);
+std::string visualize_matrix_c(
+    const legacyMatrix<std::complex<float>, size_t, Layout::ColMajor>& A);
+std::string visualize_matrix_z(
+    const legacyMatrix<std::complex<double>, size_t, Layout::ColMajor>& A);
+std::string visualize_rowmajormatrix_r(
+    const legacyMatrix<float, size_t, Layout::RowMajor>& A);
+std::string visualize_rowmajormatrix_d(
+    const legacyMatrix<double, size_t, Layout::RowMajor>& A);
+std::string visualize_rowmajormatrix_c(
+    const legacyMatrix<std::complex<float>, size_t, Layout::RowMajor>& A);
+std::string visualize_rowmajormatrix_z(
+    const legacyMatrix<std::complex<double>, size_t, Layout::RowMajor>& A);
 
 }  // namespace tlapack
