+++ test/src/CMakeLists.txt
@@ -6,29 +6,35 @@
 
 # Configurations
 include_directories("${CMAKE_CURRENT_SOURCE_DIR}/../include")
-link_libraries(Catch2::Catch2WithMain tlapack)
+list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
+
+# Common libraries
+add_library(testutils testutils.cpp)
+target_link_libraries(testutils PUBLIC Catch2::Catch2WithMain)
+target_link_libraries(testutils PUBLIC tlapack)
 
 if(TLAPACK_TEST_EIGEN)
   find_package(Eigen3 REQUIRED)
-  link_libraries(Eigen3::Eigen)
-  add_compile_definitions(TLAPACK_TEST_EIGEN)
+  target_link_libraries(testutils PUBLIC Eigen3::Eigen)
+  target_compile_definitions(testutils PUBLIC TLAPACK_TEST_EIGEN)
 endif()
 
 if(TLAPACK_TEST_MDSPAN)
   find_package(mdspan REQUIRED)
-  link_libraries(std::mdspan)
-  add_compile_definitions(TLAPACK_TEST_MDSPAN)
+  target_link_libraries(testutils PUBLIC std::mdspan)
+  target_compile_definitions(testutils PUBLIC TLAPACK_TEST_MDSPAN)
 endif()
-
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
 
 if(TLAPACK_TEST_MPFR)
   find_package(MPFR 2.3.1 REQUIRED)
   find_package(GMP 4.2.1 REQUIRED)
-  include_directories(${MPFR_INCLUDES} ${GMP_INCLUDES})
-  link_libraries(${MPFR_LIBRARIES} ${GMP_LIBRARIES})
-  add_compile_definitions(TLAPACK_TEST_MPFR)
+  target_include_directories(testutils PUBLIC ${MPFR_INCLUDES} ${GMP_INCLUDES})
+  target_link_libraries(testutils PUBLIC ${MPFR_LIBRARIES} ${GMP_LIBRARIES})
+  target_compile_definitions(testutils PUBLIC TLAPACK_TEST_MPFR)
 endif()
+
+# Use this library on all tests
+link_libraries(testutils)
 
 # Set the directory where to put testers
 set(TLAPACK_RUNTIME_OUTPUT_DIRECTORY "${CMAKE_RUNTIME_OUTPUT_DIRECTORY}")
@@ -89,5 +95,8 @@ include(Catch)
 get_property(DIRECTORY_BUILDSYSTEM_TARGETS DIRECTORY PROPERTY BUILDSYSTEM_TARGETS)
 
 foreach(target ${DIRECTORY_BUILDSYSTEM_TARGETS})
+  if(NOT target MATCHES "^test_")
+    continue()
+  endif()
   catch_discover_tests(${target})
 endforeach()
